Added files from start page branch into "my-game" directory
Added files from start page branch into "my-game" directory, renamed nathalia's page.tsx file to game.tsx.
