+++ StoryQuest/my-game/src/app/HomePage/AnimatedTitleStyles.css
@@ -1,0 +1,109 @@
+/* CSS styling for animated title in homepage */
+.animated-title {
+    font-size: 10vw;
+    line-height: 1;
+    color: #001e51;
+    font-family: 'Slackey', sans-serif;
+    text-shadow: 3px 3px 4px rgba(255, 255, 255, 1),
+    -3px -3px 4px rgba(255, 255, 255, 1),
+    3px -3px 4px rgba(255, 255, 255, 1),
+    -3px 3px 4px rgba(255, 255, 255, 1);
+
+}
+
+.animated-title > span {
+    display: inline-block;
+    overflow: hidden;
+    transition-duration: 0.2s;
+    transition-property: transform;
+    text-shadow: 2px 2px 4px rgba(255, 255, 255, 1);
+}
+
+.animated-title > span span {
+    display: inline-block;
+    transition-duration: 0.5s;
+    transition-property: transform;
+    transition-timing-function: cubic-bezier(0.34, 1.56, 0.64, 1);
+}
+
+/* Animation & Shadow for individual spans */
+.animated-title > span:nth-child(odd) {
+    transform: translateY(50%);
+}
+
+.animated-title > span:nth-child(odd) span {
+    transform: translateY(-100%);
+}
+
+.animated-title > span:nth-child(even) {
+    transform: translateY(-50%);
+}
+
+.animated-title > span:nth-child(even) span {
+    transform: translateY(100%);
+}
+
+/* Wave animation applied when "loaded" class is added */
+.animated-title.loaded > span {
+    transform: translateY(0);
+}
+
+.animated-title.loaded > span span {
+    transform: translateY(0);
+}
+
+/* Wave animation after initial animation */
+.animated-title.wave > span {
+    animation: wave 2s ease-in-out infinite;
+}
+
+/* Delay for each letter of the title */
+.animated-title.wave > span:nth-child(1) {
+    animation-delay: 0s;
+}
+
+.animated-title.wave > span:nth-child(2) {
+    animation-delay: 0.1s;
+}
+
+.animated-title.wave > span:nth-child(3) {
+    animation-delay: 0.2s;
+}
+
+.animated-title.wave > span:nth-child(4) {
+    animation-delay: 0.3s;
+}
+
+.animated-title.wave > span:nth-child(5) {
+    animation-delay: 0.4s;
+}
+
+.animated-title.wave > span:nth-child(6) {
+    animation-delay: 0.5s;
+}
+
+.animated-title.wave > span:nth-child(7) {
+    animation-delay: 0.6s;
+}
+
+.animated-title.wave > span:nth-child(8) {
+    animation-delay: 0.7s;
+}
+
+.animated-title.wave > span:nth-child(9) {
+    animation-delay: 0.8s;
+}
+
+.animated-title.wave > span:nth-child(10) {
+    animation-delay: 0.9s;
+}
+
+/* Keyframes for wave animation */
+@keyframes wave {
+    0%, 100% {
+        transform: translateY(0);
+    }
+    50% {
+        transform: translateY(-20px);
+    }
+}
