+++ StoryQuest/my-game/src/app/HomePage/HomePageStyles.css
@@ -1,0 +1,44 @@
+.page-container {
+    height: 100vh;
+    width: 100vw;
+    margin: 0;
+    padding: 0;
+}
+
+.title-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding-top: 50px;
+    padding-bottom: 50px;
+}
+
+.button-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+}
+
+.button-padding {
+    padding-top: 20px;
+    padding-bottom: 20px;
+}
+
+.button-box {
+    border: 5px solid white;
+    border-radius: 45px;
+    display: inline-block;
+    margin: 10px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+}
+
+.copyright-text {
+    font-family: 'Slackey', sans-serif;
+    text-align: right;
+    position: absolute;
+    bottom: 0;
+    right: 15px;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 1);
+}
